move accounts-related functions back to their client/server env & add the ability to customize the websocket endpoint
