+++ src/main-server.js
@@ -6,8 +6,9 @@ import { SubscriptionServer } from 'subscriptions-transport-ws';
 
 import { Meteor } from 'meteor/meteor';
 import { WebApp } from 'meteor/webapp';
+import { Accounts } from 'meteor/accounts-base';
+import { check } from 'meteor/check';
 
-import { addCurrentUserToContext } from './accounts';
 import './check-npm.js';
 
 // import the configuration functions from the client so they can be used
@@ -28,6 +29,7 @@ const defaultServerConfig = {
   graphiqlOptions: {
     passHeader: "'meteor-login-token': localStorage['Meteor.loginToken']",
   },
+  subscriptionsPath: '/subscriptions',
   subscriptionSetupFunctions: {},
   subscriptionLifecycle: {},
 };
@@ -156,8 +158,57 @@ export const createApolloServer = (customOptions = {}, customConfig = {}) => {
       {
         // bind the subscription server to Meteor WebApp
         server: WebApp.httpServer,
-        path: '/subscriptions',
+        path: config.subscriptionsPath,
       }
     );
   }
+};
+
+// take the existing context and return a new extended context with the current
+// user if relevant (i.e. valid login token)
+const addCurrentUserToContext = async (context, loginToken) => {
+  // there is a possible current user connected!
+  if (loginToken) {
+    // throw an error if the token is not a string
+    check(loginToken, String);
+
+    // the hashed token is the key to find the possible current user in the db
+    const hashedToken = Accounts._hashLoginToken(loginToken);
+
+    // get the possible current user from the database
+    // note: no need of a fiber aware findOne + a fiber aware call break tests
+    // runned with practicalmeteor:mocha if eslint is enabled
+    const currentUser = await Meteor.users.rawCollection().findOne({
+      'services.resume.loginTokens.hashedToken': hashedToken,
+    });
+
+    // the current user exists
+    if (currentUser) {
+      // find the right login token corresponding, the current user may have
+      // several sessions logged on different browsers / computers
+      const tokenInformation = currentUser.services.resume.loginTokens.find(
+        tokenInfo => tokenInfo.hashedToken === hashedToken
+      );
+
+      // get an exploitable token expiration date
+      const expiresAt = Accounts._tokenExpiration(tokenInformation.when);
+
+      // true if the token is expired
+      const isExpired = expiresAt < new Date();
+
+      // if the token is still valid, give access to the current user
+      // information in the resolvers context
+      if (!isExpired) {
+        // return a new context object with the current user & her id
+        return {
+          ...context,
+          user: currentUser,
+          userId: currentUser._id,
+        };
+      }
+    }
+  }
+
+  // return the context as passed
+  return context;
 };
